+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Next Major
 * **BC BREAK**: removed `dom/form`. Inline the implementation instead.
 * **BC BREAK**: removed `unistore`. Implement it yourself instead.
 * **BC BREAK**: removed `mount::mountJsxWithStore()`. No alternative present, you need to wrap it yourself.
+* **BC BREAK**: The class `ui/Slug` is no longer a default export but instead a named export.